src/dss-ui-app/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Deprecated==1.2.13
Flask==2.0.3
googleapis-common-protos==1.56.0
grpcio==1.44.0
idna==3.3
itsdangerous==2.1.1
jaeger-client==4.8.0
Jinja2==3.0.3
MarkupSafe==2.1.1
opentelemetry-api==1.10.0
opentelemetry-exporter-jaeger==1.10.0
opentelemetry-exporter-jaeger-proto-grpc==1.10.0
opentelemetry-exporter-jaeger-thrift==1.10.0
opentelemetry-instrumentation==0.29b0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-instrumentation-requests==0.29b0
opentelemetry-instrumentation-wsgi==0.29b0
opentelemetry-sdk==1.10.0
opentelemetry-semantic-conventions==0.29b0
opentelemetry-util-http==0.29b0
opentracing==2.4.0
protobuf==3.19.4
requests==2.27.1
six==1.16.0
threadloop==1.0.2
thrift==0.15.0
tornado==6.1
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.14.0
